backend/requirements-dev.txt
# Production dependencies
-r requirements.txt

# Development dependencies
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
isort==5.13.2
django-debug-toolbar==4.2.0
django-extensions==3.2.3
